+++ src/App.jsx
@@ -1,5 +1,4 @@
-import Index from "./components/One/index";
-import About from "./components/AboutSection/About";
+import About from "./components/AboutSection";
 
 function App() {
   return <About />;
