+++ src/components/AboutSection/stats.js
