+++ src/globals.css
@@ -2,7 +2,13 @@
 @tailwind components;
 @tailwind utilities;
 
-.custom-gradient {
+@layer components {
+    .primary-gradient {
+        @apply rounded-sm bg-gradient-to-br from-purple-600 via-blue-500 to-yellow-300;
+    }
+}
+
+/* .custom-gradient {
     border-radius: 2px;
     background: linear-gradient(135deg, #665DCD 0%, #5FA4E6 44.76%, #D2AB67 100%);
-}
+} */
